Avoid compiler warnings on 64Bit systems
git-svn-id: svn://tug.org/texlive/trunk/Build/source@18349 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/web2c/alephdir/aleph.c
@@ -179,7 +179,8 @@ new_input_line (FILE *f, halfword themode)
  if (i != EOF && (((themode!=2) && (i != '\n')) || ((themode==2) && (i != 0x25))))
     {
       (void) fprintf (stderr,
-                     "! Unable to read an entire line---ocp_buf_size=%ld.\n",ocpbufsize);
+                     "! Unable to read an entire line---ocp_buf_size=%ld.\n",
+                     (long) ocpbufsize);
       (void) fprintf (stderr, "Please increase ocp_buf_size in texmf.cnf.\n");
       uexit (1);
     }
